pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "handedness-rotation"
version = "0.0.1"
authors = [
  { name = "Ryuto Yoshida" },
]
description = "3D rotations with coordinate handedness and axis conventions; extends the rotation package."
readme = "README.md"
requires-python = ">=3.10"
classifiers = [
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
dependencies = [
    "numpy",
    "scipy",
    "cartesian-axis-3d @ git+https://github.com/ry-yoshida-private/CartesianAxis3D.git",
    "rotation @ git+https://github.com/ry-yoshida-private/Rotation.git",
]

[project.urls]
"Homepage" = "https://github.com/ry-yoshida-private/HandednessRotation"
"Bug Tracker" = "https://github.com/ry-yoshida-private/HandednessRotation/issues"

[tool.setuptools.packages.find]
where = ["src"]
